requirements.txt
streamlit
requests
pandas
numpy
pywebview
